+++ game/static/game/js/level_editor.js
@@ -313,76 +313,7 @@ ocargo.LevelEditor = function(levelId) {
                 $('#cow').click(function () {
                     new InternalCow({group: cowGroups[$('#cow_group_select').val()]});
                 });
-
-                //"Advanced" button
-                $('#advanced_cow_options_button').on('click', function () {
-                    $(this).toggleClass('cow_navigation_button_pressed');
-
-                    var showIconSrc = "/static/game/image/icons/show.svg";
-                    var hideIconSrc = "/static/game/image/icons/hide_button.svg";
-                    var advancedButtonIcon = $('#advanced_cow_options_button_icon');
-                    var iconSrc = (advancedButtonIcon.attr("src") === showIconSrc) ? hideIconSrc : showIconSrc;
-                    advancedButtonIcon.attr("src", iconSrc);
-
-                    var advancedDivHidden = $('#cow_advanced_div').css('display') === 'none';
-
-                    $('#cow_advanced_div').slideToggle(1000);
-
-                    //Only scroll down if advanced options were previously hidden
-                    if (advancedDivHidden) {
-                        $('#tab_panes_wrapper').animate({scrollTop: $('#tab_panes_wrapper')[0].scrollHeight},
-                            1500,
-                            "swing");
-                    }
-                });
-
-                // Set up cow type selector and listener
-                $('#cow_type_select').append($('<option>', {value: ocargo.Cow.WHITE})
-                    .text(gettext('White')));
-                $('#cow_type_select').append($('<option>', {value: ocargo.Cow.BROWN})
-                    .text(gettext('Brown')));
-                $('#cow_type_select').on('change', function () {
-                    var selectedGroupId = $('#cow_group_select').val();
-                    var selectedType = $('#cow_type_select').val();
-                    cowGroups[selectedGroupId].type = selectedType;
-                    var newUrl = ocargo.Drawing.raphaelImageDir + ocargo.Drawing.cowUrl(selectedType);
-                    $('#cow').attr('src', newUrl);
-
-                    for(var i = cows.length - 1; i >= 0; i--) {
-                        if(cows[i].data.group.id === selectedGroupId) {
-                            cows[i].image.attr({src: newUrl})
-                        }
-                    }
-
-                });
-
-
-                //Group select element (has to be initialised after the min and max spinners are created)
-                $('#cow_group_select').on('change', function () {
-                    $('#cow_group_select').css('background-color', cowGroups[this.value].color);
-                    //Set max values of min & max spinners
-                    var groupId = this.value;
-                    var noOfValidCowsInGroup = 0;
-                    for (var i = 0; i < cows.length; i++) {
-                        if (cows[i].valid && cows[i].data.group.id === groupId) {
-                            noOfValidCowsInGroup++;
-                        }
-                    }
-
-                    // Set cow type
-                    $('#cow_type_select').val(cowGroups[this.value].type).change();
-
-
-                });
-
-                if (Object.keys(cowGroups).length == 0) {
-                    addCowGroup();
-                }
-
-                $('#add_cow_group_button').click(addCowGroup);
-                $('#remove_cow_group_button').click(removeCowGroup);
             }
-
         }
 
         function setupCharacterTab() {